[mp] Use runtime scope for all objects
Instead of relying on the scope passed in when we create the AST node,
scope should be evaluated on runtime. A more specific use case for this
is when we add support for function evaluation, each function evaluation
will be given a new scope but the AST node will only be created once.